simx86: move Node(Un)Linker to trees.c
This way NodeUnlinker can be static.
