+++ src/base/emu-i386/simx86/trees.h
@@ -161,6 +161,7 @@ void InitTrees(void);
 unsigned int FindPC(const unsigned char *addr);
 int InvalidateNodeRange(int addr, int len, unsigned char *eip);
 void avltr_delete(const int key);
+void NodeLinker(TNode *LG, TNode *G);
 
 #ifdef DEBUG_TREE
 extern FILE *tLog;
